+++ hudi-common/src/test/java/org/apache/hudi/common/table/read/lsm/TestLsmFileIterators.java
@@ -22,9 +22,11 @@ package org.apache.hudi.common.table.read.lsm;
 import org.apache.hudi.common.engine.HoodieReaderContext;
 import org.apache.hudi.common.engine.RecordContext;
 import org.apache.hudi.common.model.HoodieBaseFile;
+import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.schema.HoodieSchema;
 import org.apache.hudi.common.schema.HoodieSchemaField;
 import org.apache.hudi.common.schema.HoodieSchemaType;
+import org.apache.hudi.common.schema.HoodieSchemas;
 import org.apache.hudi.common.table.read.BufferedRecord;
 import org.apache.hudi.common.table.read.DeleteContext;
 import org.apache.hudi.common.table.read.FileGroupReaderSchemaHandler;
@@ -38,11 +40,14 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
@@ -51,6 +56,26 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 class TestLsmFileIterators {
+
+  @Test
+  void testNativeDeleteRecordPreservesOrderingValue() {
+    HoodieReaderContext<Map<String, Object>> readerContext = mock(HoodieReaderContext.class);
+    RecordContext<Map<String, Object>> recordContext = mock(RecordContext.class);
+    Map<String, Object> record = Collections.emptyMap();
+    List<String> orderingFields = Collections.singletonList("ts");
+    HoodieSchema deleteLogSchema = HoodieSchemas.createDeleteLogSchema(tableSchema(), orderingFields);
+
+    when(readerContext.getRecordContext()).thenReturn(recordContext);
+    when(recordContext.getValue(record, deleteLogSchema, HoodieRecord.RECORD_KEY_METADATA_FIELD)).thenReturn("key1");
+    when(recordContext.getOrderingValue(eq(record), eq(deleteLogSchema), eq(orderingFields))).thenReturn(42L);
+
+    BufferedRecord<Map<String, Object>> deleteRecord =
+        LsmFileIterators.createNativeDeleteRecord(readerContext, record, deleteLogSchema, orderingFields);
+
+    assertEquals("key1", deleteRecord.getRecordKey());
+    assertEquals(42L, deleteRecord.getOrderingValue());
+    assertTrue(deleteRecord.isDelete());
+  }
 
   @Test
   void testCreateBaseFileIteratorUsesIteratorMode() throws Exception {
@@ -130,5 +155,11 @@ class TestLsmFileIterators {
 
     iterator.close();
     verify(fileIterator).close();
+  }
+
+  private static HoodieSchema tableSchema() {
+    return HoodieSchema.createRecord("test_record", null, null, Arrays.asList(
+        HoodieSchemaField.of("id", HoodieSchema.create(HoodieSchemaType.STRING)),
+        HoodieSchemaField.of("ts", HoodieSchema.create(HoodieSchemaType.LONG))));
   }
 }
